requirements_production.txt

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1

# Telegram
python-telegram-bot==20.7

# Trading & Analysis
ccxt==4.1.49
pandas==2.1.4
numpy==1.26.2
plotly==5.17.0
ta-lib==0.4.28

# Payment
stripe==7.8.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Security
cryptography==41.0.8
pyjwt==2.8.0

# Performance
asyncio==3.4.3
aiohttp==3.9.1
celery==5.3.4

# Database
asyncpg==0.29.0

# Development (remover em produção)
pytest==7.4.3
pytest-asyncio==0.21.1
